separate main window to MainWindow.py; and small cleanup
