+++ main.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env py -3-32
 # -*- coding: utf-8 -*-
 
-from PyQt5.QtWidgets import QApplication
 import sys
-import mTwit.ui as ui
+from mTwit.MainWindow import MainWindow
+from PyQt5.QtWidgets import QApplication
 
 # api = Auth.api  # Twitter Rest API Object (Option GUI Required)
 # auth = Auth.auth # Tweepy Auth Object (Option GUI Required)
@@ -12,5 +12,5 @@ app = QApplication(sys.argv)
 app.setStyleSheet(
     "QMainWindow{background-image: url('image/window.png');border: 0px solid black};")
 
-window = ui.mainWindow()
+window = MainWindow()
 sys.exit(app.exec_())
